Validators Functionality Added for Each Actor - Devanshu Garg

+++ CRS-JEDI-7-Flipkart-POS/src/com/crs/flipkart/business/AdminService.java
@@ -23,6 +23,7 @@ import com.crs.flipkart.exceptions.ProfessorNotFoundException;
 import com.crs.flipkart.exceptions.StudentNotFoundForApprovalException;
 import com.crs.flipkart.exceptions.UserIdAlreadyInUseException;
 import com.crs.flipkart.utils.Utils;
+import com.crs.flipkart.validator.AdminValidator;
 
 /**
  * @author LENOVO
@@ -58,10 +59,6 @@ public class AdminService implements AdminInterface {
 	AdminDaoInterface adminDaoOperation = AdminDaoOperation.getInstance();
 	
 	Scanner sc = new Scanner(System.in);
-
-    Vector<Student> StudentList = new Vector<>();
-    Vector<Professor> ProfessorList = new Vector<>();
-    Vector<Course> CourseList = new Vector<>();
 
     @Override
 	public void addProfessor(Professor professor) throws ProfessorNotAddedException, UserIdAlreadyInUseException {
@@ -102,18 +99,9 @@ public class AdminService implements AdminInterface {
 	@Override
 	public void approveStudentRegistration(int studentId, Vector<Student> pendingStudents) throws StudentNotFoundForApprovalException {
 		
-		Boolean isValidUnapprovedStudent = false;
- 		for(Student student : pendingStudents) {
- 			if(studentId == student.getStudentId()) {
- 				isValidUnapprovedStudent = true;
- 				break;
- 			}
- 		}
-
- 		if(!isValidUnapprovedStudent) {
- 			System.out.println("Student not found.");
- 			return;
- 		}
+		if (!AdminValidator.isValidUnapprovedStudent(studentId, pendingStudents)) {
+			throw new StudentNotFoundForApprovalException(studentId);
+		}
  		
  		try {
  			adminDaoOperation.approveStudentRegistration(studentId);
@@ -150,8 +138,13 @@ public class AdminService implements AdminInterface {
 	}
 
 	@Override
-	public void addCourse(Course course) throws CourseAlreadyExistsException {
+	public void addCourse(Course course, Vector<Course> courseList) throws CourseAlreadyExistsException {
 
+		if (!AdminValidator.isValidNewCourse(course, courseList)) {
+			logger.error("Course with Course Code: " + course.getCourseId() + " is already present in Catalog.");
+			throw new CourseAlreadyExistsException(course.getCourseId());
+		}
+		
 		try {
 			adminDaoOperation.addCourse(course);
 		} catch (CourseAlreadyExistsException e) {
@@ -166,8 +159,13 @@ public class AdminService implements AdminInterface {
 	}
 
 	@Override
-	public void deleteCourse(int courseId) throws CourseNotFoundException, CourseNotDeletedException {
-
+	public void deleteCourse(int courseId, Vector<Course> courseList) throws CourseNotFoundException, CourseNotDeletedException {
+		
+		if (!AdminValidator.isValidDropCourse(courseId, courseList)) {
+			logger.error("Course with Course Code: " + courseId + " is not present in Catalog.");
+			throw new CourseNotFoundException(courseId);
+		}
+		
 		try {
 			adminDaoOperation.deleteCourse(courseId);
 		} catch (CourseNotFoundException e) {
